Redirect user to login route if necessary
We need the user’s credentials to load a recent unpatrolled change, so
redirect them to the login page if they aren’t logged in yet.

Note that we don’t decorate any other routes like this so far. One
particular diff page should be possible to display without credentials –
they just won’t be able to do anything with it.