Fix user_rights if user is not logged in
If the user is not logged in yet, report that we have no rights, instead
of crashing. This automatically disables the “patrol” and “rollback”
buttons because the user will not have the respective rights, which is
fine. (If they still decide to POST to those routes manually, the
resulting crash is acceptable, I think.)