Replace MobileDiff embedding
Instead of embedding Special:MobileDiff in an iframe, which is
problematic (if the user is logged in on m.wikidata.org, there will be a
“mark as patrolled” link and therefore MediaWiki will send an
X-Frame-Options: deny header to protect against clickjacking attacks),
render the diff via the API and include that directly on the page, then
embed the regular page view beneath that.

To make the diff look good, we add a surrounding table, based on
MediaWiki’s own diff pages, and add some of MediaWiki’s CSS via two
ResourceLoader modules as well.

For the regular page view, we use the Minerva Neue skin, but in desktop
format, not mobile format. Minerva Neue, the default skin for mobile
format, gives us a fairly clean, decluttered page, but in mobile format
Wikibase emits statements in a different format which I don’t think we
want here. (But I might reevaluate this later.) The height of this
iframe is demoted to 60% of the viewport to account for the fact that
there’s now some extra content above the iframe (the diff), but that’s
not enough to keep the buttons in view on mobile devices; however,
cleaning up the layout is left to a separate commit.