Don’t select new page creations
The tool would have to work a bit differently to support revisions that
create new pages: there is no older revision to compare to (which
manifests itself as a MediaWiki API warning and a KeyError on
results['fromuser']), and rollback doesn’t really work, you would have
to file a deletion request instead. For now, let’s avoid those issues by
only asking the recentchanges API for proper edits – if a user manually
navigates to /diff/<id of creation>, then they’ll get a server error and
that’s okay in my opinion (as long as it doesn’t bring down the whole
server, which it doesn’t).