Remove Flexbox layout again
On mobile browsers, the viewport height includes the browser chrome
(address bar etc.), so a height: 100% page actually needs a bit of
scrolling to be fully visible. Also, with the buttons stacked vertically
the iframe has so little space that it’s almost completely useless.

Instead, make the form containing the buttons stick to the bottom of the
page, so that they can always be used, and give the iframe a fixed
portion of the viewport height again. On large displays, this means that
(almost) no scrolling is necessary on the page; on small displays, on
the other hand, you can choose between mostly seeing the diff and mostly
seeing the iframe, with the buttons always being available. I think this
works out much better in practice.