Fix relative URLs in diffs
Wikibase uses relative URLs when formatting links to diffs, which
doesn’t really work when we embed the resulting HTML in our own tool,
under a different domain. Eventually we’ll hopefully fix this on the
Wikibase end, but for now we can work around it by post-processing the
HTML a bit. (I also intend to use the Beautiful Soup library for another
thing anyways, so importing it for this doesn’t hurt.)