Add unicodescripts module
This module, which is built from the Unicode source data file by the
make-unicodescripts.py script, returns the script of a given Unicode
character. (This data is not available in Python’s own unicodedata
module.) This will be used to hide diffs in scripts the user can’t read.

Even though unicodescripts.py is a generated file, I’m choosing to
commit it because:

- it’s not overly long,
- it only changes very rarely, and
- this means you don’t need an extra build script to use the tool.

When a new version of the Unicode standard is released, the module may
be rebuilt using the following command:

    make -B unicodescripts.py

The current version corresponds to Unicode 11.0.0.