Add function to get primary script of diff
This function analyzes the language-specific parts of a Wikibase diff
HTML (excluding e. g. entity labels) and returns their most common
script (apart from the 'Common' script – punctuation etc.). This will be
used to skip diffs in scripts the user can’t read.

If the diff touches both language-specific and language-agnostic parts
(e. g. it changes a statement value but also a label at the same time),
then the script of the language-specific part will be returned no matter
the proportion of language-specific to -agnostic parts. That should be
rare, though, so I think it’s acceptable.