Skip diffs in scripts the user can’t read
If there is a list of “supported” scripts in the user’s session, analyze
the diffs of unpatrolled edits and only rediret the user to those they
can read. This makes any_diff() somewhat slower, but it seems to be okay
in practice. (Note that diffs with no primary script, e. g. for edits
touching statements with entity values, will be shown to any user.)

There is no way for the user to set this list of supported scripts yet –
that will be added next.