Also ignore Inherited and Unknown scripts
Inherited is the script of diacritics and other kinds of characters that
can be combined with characters from multiple other scripts and inherit
their script. Unknown is the script of characters not known to the
unicodescripts database (shouldn’t happen, but who knows). Both are not
useful to return as the primary script of a text.