Infer user’s supported scripts from Babel
We get the languages a user understands from their Babel, ask MediaWiki
for the autonyms of those languages (the names of the languages in that
language), take the scripts of those languages, and use those to
populate the selection of scripts on the settings page if the user never
set the supported languages before.

Note that, until the user saves the settings page (using this selection
or any other selection they want), any_diff() doesn’t filter by script,
so this is not used to filter by default, just to prevent the user with
a reasonable default on the settings page. We also add Latin to this
default selection regardless of whether it’s actually present in the
user’s Babel or not. (However, if they remove it from the selection
before submitting the form, we respect that change.)