Rename skipped_ids to skipped_rev_ids
We’ll soon start to track some page IDs as well, so let’s make this name
slightly less ambiguous.

This means that the skipped_ids of the user will just be discarded,
which is okay because the tool isn’t publicly used yet. (That is, some
kind of migration mechanism is just not necessary.)