Automatically skip pages if user skips multiple revisions
If the user has skipped more than one revision of a page without ever
having otherwise acted on revisions of that page (i. e. patrolled or
rolled back any of them), then we assume that there’s a series of edits
here which the user can’t patrol, and automatically skip all revisions
of that page in the future.

This uses the same expiry mechanism as for revision IDs, though the
sorting behavior doesn’t make much sense here; I’ll fix that later.