Catch errors on rollback
Rollback can fail for a variety of reasons; let’s catch errors and show
the message to the user.

Since the mwapi library doesn’t support non-legacy error formats yet, we
have to do an extra API call to turn the wikitext API error into HTML we
can show to the user.