Update ID list management
Move the automatic truncation from get() to append() – during normal use
of the tool, each append() will correspond to at least one get() (from
the next navigation to the any_diff() endpoint), but get()ting the list
multiple times doesn’t mean it needs to be truncated multiple times, so
doing that in append() makes much more sense.

Also, don’t sort the list by the numeric ID value – that makes sense for
revision IDs, but not so much for page IDs. Instead, just treat it as a
queue, inserting new IDs at the front and trimming the list at the back.
(We do it this way around, instead of appending at the back, because
checking if an ID is a member of the list goes from front to back, and
at least for revision IDs search results are much more likely to be
found among the recent IDs, so we want those to be at the front of the
list, where they’re found faster.)