Reduce number of page IDs stored in the session
The long lists of revision and page IDs only appeared to fit in the
session cookie in the test because the test actually added the same page
IDs to all three page ID lists, which meant they compressed really well.
Fixing this reveals that completely full lists of page IDs don’t
actually fit in the session cookie, so let’s further lower the limit for
page IDs (which we detect based on the name).