Make Flask session permanent
According to its documentation, the Flask session expires with the
browser session by default. I haven’t experienced it expiring quite that
frequently in practice, but it does expire from time to time (especially
on mobile, it seems), which is annoying – not so much because of the
lost OAuth credentials (reauthenticating is quick), but more because of
the lost settings. Let’s instruct Flask to keep the session around
permanently (for 31 days by default).