Redirect to index after saving settings
Re-displaying the settings page isn’t very useful, so let’s redirect the
user to the index page instead (using 303 See Other, to ensure that the
client doesn’t repeat the POST request).

It would be even more useful to send the user directly to the any_diff()
route, but I think that would also be a bit too confusing. The index
page should be fine.