Catch permission errors when getting unpatrolled revisions
Users who are not autoconfirmed are not allowed to retrieve unpatrolled
recent changes; previously, this would result in an internal server
error. We now catch this error and present it to the user, with the most
likely cause.

The error page is more vague than it needs to be: if we wanted to, we
could get the user’s groups and rights, and check if they are
autoconfirmed and/or do or don’t have the required rights. But that just
raises the question of what error to show if it’s not the expected case
of “user is not autoconfirmed”. Since I’m not aware of other cases where
this would happen, I prefer to keep the error page vague like this.

(We have to wrap the whole loop in a try/catch – unpatrolled_changes()
is a generator, and any exception it raises is only thrown once we start
to iterate it, extracting the unpatrolled_changes() call into a variable
and wrapping just that doesn’t work.)

Thanks to blinry for finding this bug and testing the fix!
https://twitter.com/blinry/status/1111590362349477889