Analyze scripts from English-language diffs
primary_script_of_diff() uses the diff headers to decide which parts of
the diff to analyze and which to skip. This requires the diff headers to
be in English, otherwise they can’t be analyzed. Previously, they would
be in the user language, so users with a non-English user language would
sometimes get a diff that the function couldn’t analyze, and the tool
would fall back to showing them any diff regardless of script. Fix that
by requesting an English diff for analysis. (Note that this is not the
same as the diff we actually show to the user, which is still in the
user language.)

Thanks to User:Sjoerddebruin for reporting and helping to track down
this bug!