Be more forgiving of CSRF errors on “skip”
I sometimes encounter these when there are connection hiccups on the
mobile network – the first click on “skip” will go through to the server
but the response will get lost, and then the second click will be a CSRF
error. It’s perfectly fine to just redirect to the next diff (without
marking the current one as skipped) in such an event.