Upgrade dependencies
    pip-compile -U

Werkzeug 2.3 changes the way cookies are handled, which supposedly
improves performance significantly, but also broke our test for ensuring
the session fits in a cookie. I wasted a bunch of time trying to figure
out how to get a proper Cookie header or CookieJar out of a new-style
test session, but didn’t get anywhere; in the end I decided that
hard-coding a cookie header value with just one cookie (session cookie)
is probably good enough.