Make session permanent after login
By default, the Flask session only survives until the user closes the
browser [1], which is annoying. With this, the session will survive for
31 days; as far as I can tell, that expiration time is refreshed each
time the user visits the tool, so in practice it can last much longer as
long as the user uses the tool regularly.

In logout(), flask.session.clear() implicitly resets the permanent flag
(it’s just stored in the dict, under the key “_permanent”, so clearing
the dict includes the flag), so we don’t need to reset it explicitly
(like we do in some other tools).

[1]: https://flask.palletsprojects.com/en/2.3.x/api/#flask.session.permanent